fix: auth types
