+++ src/app/(auth)/signup/page.tsx
@@ -12,14 +12,8 @@ import PasswordInput from '@/app/components/Auth/Form/PasswordInput';
 import AuthHeader from '@/app/components/Auth/Header';
 import LayerSpinner from '@/app/components/Spinner/LayerSpinner';
 import useSignup from '@/app/hooks/auth/useSignup';
+import { SignUpInput } from '@/app/types/form/auth';
 
-interface SignUpInput {
-  email: string;
-  name: string;
-  nickname: string;
-  password: string;
-  passwordConfirmation: string;
-}
 
 export default function Signup() {
   const { mutate, isPending } = useSignup();
